refactor: index.tsのルート登録をファクトリ関数に分割 (#136)
- registerRoutes(app)をapps/backend/src/routes/index.tsに分離
- registerMiddleware(app)をapps/backend/src/middleware/index.tsに分離
- scheduledハンドラをapps/backend/src/scheduled.tsに分離
- index.tsを17行のエントリポイントに圧縮
- 開発ガイドにルート追加手順を更新
