+++ docs/基本/開発ガイド.md
@@ -166,13 +166,23 @@ newEndpoint.get('/', async (c) => {
 export default newEndpoint;
 ```
 
-`apps/backend/src/index.ts`で統合：
+`apps/backend/src/routes/index.ts`の`registerRoutes`関数に追加：
 
 ```typescript
-import newEndpoint from './routes/new-endpoint';
+// apps/backend/src/routes/index.ts
+import newEndpoint from './new-endpoint';
 
-app.route('/api/new-endpoint', newEndpoint);
+export function registerRoutes(app: Hono<AppEnv>): void {
+  // ... 既存のルート ...
+  app.route('/api/new-endpoint', newEndpoint);
+}
 ```
+
+**ルート追加手順:**
+1. `apps/backend/src/routes/new-endpoint.ts` にルートファイルを作成
+2. `apps/backend/src/routes/index.ts` の `registerRoutes` 関数にimportと`app.route()`を追加
+3. ミドルウェアが必要な場合は `apps/backend/src/middleware/index.ts` の `registerMiddleware` 関数を編集
+4. `apps/backend/src/index.ts` は変更不要（エントリポイントのみ）
 
 #### 共通クエリの追加
